move intersection screens into separate folder
